renamed resources folder
spotless checks

spotless checks

spotless checks

corrected sql file

moved tests to different files

moved tests to different files

cleaned up sql files

added attribute_value table rows

removed date checks

spotless checks

removed extra table

added gcs path

added logger
