+++ v2/sourcedb-to-spanner/src/test/resources/SchemaMapperIT/company-session-with-transformation.json
