+++ v2/sourcedb-to-spanner/src/test/resources/SingleShardWithTransformation/session.json
@@ -1,0 +1,224 @@
+{
+  "SessionName": "NewSession",
+  "EditorName": "",
+  "DatabaseType": "mysql",
+  "DatabaseName": "singleShardIT",
+  "Dialect": "google_standard_sql",
+  "Notes": null,
+  "Tags": null,
+  "SpSchema": {
+    "t171": {
+      "Name": "SingleShardWithTransformationTable",
+      "ColIds": [
+        "c172",
+        "c173",
+        "c174",
+        "c175"
+      ],
+      "ShardIdColumn": "c175",
+      "ColDefs": {
+        "c172": {
+          "Name": "pkid",
+          "T": {
+            "Name": "INT64",
+            "Len": 0,
+            "IsArray": false
+          },
+          "NotNull": true,
+          "Comment": "From: pkid int(10)",
+          "Id": "c172",
+          "AutoGen": {
+            "Name": "",
+            "GenerationType": ""
+          }
+        },
+        "c173": {
+          "Name": "name",
+          "T": {
+            "Name": "STRING",
+            "Len": 20,
+            "IsArray": false
+          },
+          "NotNull": false,
+          "Comment": "From: name varchar(20)",
+          "Id": "c173",
+          "AutoGen": {
+            "Name": "",
+            "GenerationType": ""
+          }
+        },
+        "c174": {
+          "Name": "status",
+          "T": {
+            "Name": "STRING",
+            "Len": 20,
+            "IsArray": false
+          },
+          "NotNull": false,
+          "Comment": "From: status varchar(20)",
+          "Id": "c174",
+          "AutoGen": {
+            "Name": "",
+            "GenerationType": ""
+          }
+        },
+        "c175": {
+          "Name": "migration_shard_id",
+          "T": {
+            "Name": "STRING",
+            "Len": 50,
+            "IsArray": false
+          },
+          "NotNull": false,
+          "Comment": "",
+          "Id": "c175",
+          "AutoGen": {
+            "Name": "",
+            "GenerationType": ""
+          }
+        }
+      },
+      "PrimaryKeys": [
+        {
+          "ColId": "c172",
+          "Desc": false,
+          "Order": 2
+        },
+        {
+          "ColId": "c175",
+          "Desc": false,
+          "Order": 1
+        }
+      ],
+      "ForeignKeys": null,
+      "Indexes": null,
+      "ParentId": "",
+      "Comment": "Spanner schema for source table SingleShardWithTransformationTable",
+      "Id": "t171"
+    }
+  },
+  "SyntheticPKeys": {},
+  "SrcSchema": {
+    "t171": {
+      "Name": "SingleShardWithTransformationTable",
+      "Schema": "singleShardIT",
+      "ColIds": [
+        "c172",
+        "c173",
+        "c174"
+      ],
+      "ColDefs": {
+        "c172": {
+          "Name": "pkid",
+          "Type": {
+            "Name": "int",
+            "Mods": [
+              10
+            ],
+            "ArrayBounds": null
+          },
+          "NotNull": true,
+          "Ignored": {
+            "Check": false,
+            "Identity": false,
+            "Default": false,
+            "Exclusion": false,
+            "ForeignKey": false,
+            "AutoIncrement": true
+          },
+          "Id": "c172"
+        },
+        "c173": {
+          "Name": "name",
+          "Type": {
+            "Name": "varchar",
+            "Mods": [
+              20
+            ],
+            "ArrayBounds": null
+          },
+          "NotNull": false,
+          "Ignored": {
+            "Check": false,
+            "Identity": false,
+            "Default": false,
+            "Exclusion": false,
+            "ForeignKey": false,
+            "AutoIncrement": false
+          },
+          "Id": "c173"
+        },
+        "c174": {
+          "Name": "status",
+          "Type": {
+            "Name": "varchar",
+            "Mods": [
+              20
+            ],
+            "ArrayBounds": null
+          },
+          "NotNull": false,
+          "Ignored": {
+            "Check": false,
+            "Identity": false,
+            "Default": false,
+            "Exclusion": false,
+            "ForeignKey": false,
+            "AutoIncrement": false
+          },
+          "Id": "c174"
+        }
+      },
+      "PrimaryKeys": [
+        {
+          "ColId": "c172",
+          "Desc": false,
+          "Order": 1
+        }
+      ],
+      "ForeignKeys": null,
+      "Indexes": null,
+      "Id": "t171"
+    }
+  },
+  "SchemaIssues": {
+    "t171": {
+      "ColumnLevelIssues": {
+        "c172": [
+          14,
+          11
+        ],
+        "c175": [
+          29
+        ]
+      },
+      "TableLevelIssues": null
+    }
+  },
+  "Location": {},
+  "TimezoneOffset": "+00:00",
+  "SpDialect": "google_standard_sql",
+  "UniquePKey": {},
+  "Rules": [
+    {
+      "Id": "r176",
+      "Name": "r176",
+      "Type": "add_shard_id_primary_key",
+      "ObjectType": "",
+      "AssociatedObjects": "All Tables",
+      "Enabled": true,
+      "Data": {
+        "AddedAtTheStart": true
+      },
+      "AddedOn": {
+        "TimeOffset": null
+      }
+    }
+  ],
+  "IsSharded": true,
+  "SpRegion": "",
+  "ResourceValidation": false,
+  "UI": false,
+  "SpSequences": {},
+  "SrcSequences": {}
+}
