+++ v2/sourcedb-to-spanner/src/test/resources/SingleShardWithTransformation/spanner-schema.sql
@@ -1,0 +1,6 @@
+CREATE TABLE SingleShardWithTransformationTable (
+	pkid INT64 NOT NULL,
+	name STRING(20),
+	status STRING(20),
+	migration_shard_id STRING(50),
+) PRIMARY KEY (migration_shard_id, pkid);
